MANTA-4819 Make schema-manager use cueball (#22)
* Refactor the schema-manager helper functions and move them to the utils
crate. Update the rpc_handler tests to exercise this code by using the helper
functions to create the vnode schemas in the temporary database used by the
tests.
* Use cueball and manatee primary resolver in the schema manager so that the
schema manager can automatically determine the manatee peer that is the
primary. This also involves changing the schema-manager and its helper functions
to use the postgres crate instead of diesel for communications with postgres.