MANTA-4834 schema-manager must use the default postgres user and database initially
Use one connection to the default postgres database using the default postgres
role to create the application role and database and a different connection to
the application database to create the schemas within that database. Also only
create a single connection in the connection pools since there will only ever be
a single connection needed from each one.