MANTA-4320 Would like boray functions for GC (#26)
Expose two RPC functions to support garbage collection: getgcbatch and deletegcbatch

Implement a system to handle versioned database migrations. There are two tiers
of migrations. The first is for the public schema in postgres. These are
represented as SQL files. The second tier is migrations that apply to each
individual vnode schema. The migrations are represented as templates that can be
rendered and run for each vnode on a particular physical node or shard.

Implement the migrations necessary to support the garbage
collection functionality.